// web/src/types/WebQuery.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Interval } from "./Interval";
import type { PropertyFilter } from "./PropertyFilter";

/**
 * Query string of both web endpoints:
 * `date_from`, `date_to` (as `DateRange`), `interval`, and `properties`
 * (URL-encoded JSON array of `PropertyFilter`).
 */
export type WebQuery = { date_from: string, date_to: string | null, interval: Interval | null, properties: Array<PropertyFilter>, };